pyproject.toml
[project]
name = "mvp-transformation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "taichi>=1.7.4",
]
